tbot/requirements.txt
python-telegram-bot==20.3
requests==2.32.3
python-decouple==3.8